@arnoldlankamp 2023-06-17
I'm pretty sure the main gain here is from the removal of the bounds-check-elimination workaround (i.e. the weird `while` construct that you turned back into an `if`).
This was something that improved performance in earlier versions of JDK 1.6, since the JIT compiler couldn't / wouldn't eliminate the unnecessary bounds-check on the array access in the `add` method, since it's situated after a call to a native function (`System.arraycopy`) on the execution path where the `enlarge` method was called, of which the JIT compiler has no internal knowledge, since the it can only 'see' method local code (including inlined code, but can't look into native calls), so could not guarantee the array access could never cause an out-of-bounds exception in this case, so it remained in place.
If I recall, this shortcoming was resolved in one of the later patch versions of JDK 1.6. They probably added some special cases for things like `System.arraycopy` or implemented a code path specific optimization where both possible execution paths are optimized separately, since being able to do these kinds of optimizations would also greatly benefit the performance of the data structures in the standard library.

The use of `Arrays.copyOf` will most definitely have no impact (or a very minor negative one), since it's implementation just invokes `System.arraycopy` for any non-zero length array. So all this specific change does is add a superfluous method call.

`Arrays.copyOf` is in essence nothing more than a convenience wrapper for `System.arraycopy`, which returns a value, so users can chain off of it 😉 .