@DavyLandman 2023-06-18
I was also curious. And it's the other way around. The ifcheck is actually still good. It's all in the implementation of `Arrays.copyOf`.

After only using the the arrays.copyof, it actuall was even a bit faster.

main branch:

```
Start parsing
Warmup took: 9544 ms
Warmup took: 8108 ms
Parsing took: 8043 ms
Parsing took: 8054 ms
Parsing took: 8062 ms
Parsing took: 8006 ms
Parsing took: 8077 ms
Parsing took: 8143 ms
Parsing took: 8078 ms
Parsing took: 8016 ms
Parsing took: 8103 ms
Parsing took: 8085 ms
Done: 26130420, average time: 8067 ms
```


this PR:
```
Warmup took: 8748 ms
Warmup took: 7055 ms
Parsing took: 6980 ms
Parsing took: 6967 ms
Parsing took: 7063 ms
Parsing took: 6919 ms
Parsing took: 6920 ms
Parsing took: 7011 ms
Parsing took: 6931 ms
Parsing took: 7012 ms
Parsing took: 7049 ms
Parsing took: 6935 ms
Done: 26130420, average time: 6979 ms
```

So roughly 13%.


Why is Arrays.copyOf faster? Since it's marked as a `HotSpotIntrinsicCandidate`

> The @HotSpotIntrinsicCandidate annotation is specific to the HotSpot Virtual Machine. It indicates that an annotated method may be (but is not guaranteed to be) intrinsified by the HotSpot VM. A method is intrinsified if the HotSpot VM replaces the annotated method with hand-written assembly and/or hand-written compiler IR -- a compiler intrinsic -- to improve performance. 

So I'm guessing it does the allocation and moving in a more compact way.