@arnoldlankamp 2023-06-18
Interesting, so what seems to have happened is that they basically added a second native array copy implementation. The big difference is that the new one uses SIMD.

But why they shoehorned this optimized version into the `Arrays` wrapper implementation, instead of just optimizing `System.arraycopy`, so everyone benefits, is a bit of a mystery 🤔 .