@arnoldlankamp 2023-06-18
Maybe a next step would be to replace all other occurrences of `System.arraycopy`, since `Arrays.copyOf` is currently the faster implementation?