@DavyLandman 2023-06-18
It's because with System.arrayCopy you have no guarantees about non
overlapping source & dest.

So if you know the memory cannot overlap, you can be faster.

On Sun, 18 Jun 2023, 10:38 arnoldlankamp, ***@***.***> wrote:

> Interesting, so what seems to have happened is that they basically added a
> second native array copy implementation. The big difference is that the new
> one uses SIMD.
>
> But why they shoehorned this optimized version into the Arrays wrapper
> implementation, instead of just optimizing System.arraycopy, so everyone
> benefits, is a bit of a mystery 🤔 .
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/usethesource/rascal/pull/1813#issuecomment-1596035891>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AABL3EZ5SJPSLDKNGJPTOOTXL246RANCNFSM6AAAAAAZKN7OVQ>
> .
> You are receiving this because you were mentioned.Message ID:
> ***@***.***>
>