@arnoldlankamp 2023-06-18
Yep, they'd need an `if` statement to check that first 😄 .