@jurgenvinju 2023-06-19
I'm very happy you guys made the parser more efficient.

Now I think it's time we close this and move back to usethesource/rascal-language-servers#267, because we should focus on not having the start the entire bootstrapping of the parser generator in deploy-mode _at all_ for the interpreted modules of DSLs.